Stage

+++ src/TD.Web/TD.Web.Admin/TD.Web.Admin.Contracts/Requests/KomercijalnoWebProductLinks/KomercijalnoWebProductLinksSaveRequest.cs
@@ -4,7 +4,7 @@ namespace TD.Web.Admin.Contracts.Requests.KomercijalnoWebProductLinks
 {
     public class KomercijalnoWebProductLinksSaveRequest : LSCoreSaveRequest
     {
-        public long RobaId { get; set; }
+        public int RobaId { get; set; }
         public long WebId { get; set; }
     }
 }
